fix: handle ssb reserved words in test configs
